+++ BankPrikoloff/Controllers/FileController.cs
@@ -1,4 +1,5 @@
 ﻿using BankPrikoloff.Contracts;
+using BusinessLogic.Authorization;
 using BusinessLogic.Interfaces;
 using BusinessLogic.Servises;
 using Domain.Models;
@@ -8,9 +9,10 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace BankPrikoloff.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
-    public class FileController : ControllerBase
+    public class FileController : BaseController
     {
 
         private IFileService _fileService;
@@ -22,6 +24,7 @@ namespace BankPrikoloff.Controllers
         /// <summary>
         /// Получение всех файлов
         /// </summary>
+        [Authorize(2)]
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
@@ -31,10 +34,15 @@ namespace BankPrikoloff.Controllers
         /// <summary>
         /// Получение файла по ID
         /// </summary>
+        
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(string id)
         {
             var dto = await _fileService.GetById(id);
+            if (dto.ClientId != User.ClientId && User.RoleId != 2)
+            {
+                return Unauthorized(new { message = "Unathorized" });
+            }
             return Ok(dto.Adapt<GetFileRequest>());
         }
         /// <summary>
@@ -55,6 +63,10 @@ namespace BankPrikoloff.Controllers
             var Dto = request.Adapt<Domain.Models.File>();
             Dto.ClientId = "qwerty";
             Dto.FilePath = $"File{Dto.FileId}";
+            if (Dto.ClientId != User.ClientId && User.RoleId != 2)
+            {
+                return Unauthorized(new { message = "Unathorized" });
+            }
             await _fileService.Create(Dto);
             return Ok();
         }
@@ -78,7 +90,11 @@ namespace BankPrikoloff.Controllers
         public async Task<IActionResult> Update(GetFileRequest request)
         {
             var Dto = request.Adapt<Domain.Models.File>();
-            await _fileService.Create(Dto);
+            if (Dto.ClientId != User.ClientId && User.RoleId != 2)
+            {
+                return Unauthorized(new { message = "Unathorized" });
+            }
+            await _fileService.Update(Dto);
             return Ok();
         }
         /// <summary>
@@ -87,6 +103,11 @@ namespace BankPrikoloff.Controllers
         [HttpDelete]
         public async Task<IActionResult> Delete(string id)
         {
+            var file = await _fileService.GetById(id);
+            if (file.ClientId != User.ClientId && User.RoleId != 2)
+            {
+                return Unauthorized(new { message = "Unathorized" });
+            }
             await _fileService.Delete(id);
             return Ok();
         }
